requirements.txt
alembic==0.9.1
altgraph==0.10.2
aniso8601==1.2.0
aspy.yaml==0.2.2
bdist-mpkg==0.5.0
bonjour-py==0.3
cached-property==1.3.0
click==6.7
decorator==4.0.9
Flask==0.12
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.5.1
macholib==1.5.1
Mako==1.0.6
MarkupSafe==0.23
matplotlib==1.3.1
modulegraph==0.10.4
MySQL-python==1.2.5
networkx==1.11
nltk==3.2.1
nodeenv==0.13.6
numpy==1.8.0rc1
pandas==0.19.0
pre-commit==0.8.1
py2app==0.7.3
PyJWT==1.4.2
PyMySQL==0.7.10
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pyterminalsize==0.1.0
python-dateutil==1.5
python-editor==1.0.3
pytz==2013.7
PyYAML==3.11
scikit-learn==0.17.1
scipy==0.13.0b1
six==1.4.1
sklearn==0.0
SQLAlchemy==1.1.6
textblob==0.11.1
virtualenv==15.0.1
Werkzeug==0.11.15
xattr==0.6.4
zope.interface==4.1.1
